leds: lp5562: BLN: remove reprogramming of VK blink if in blinking
Adding "blink_running" variable to  follow when the blink program
is already set to on for the Buttons. In such cases, avoid restarting
the blinking, only maintain the unblink alar function timing push
out as it was before, so that it will continue to blink.
Blink_running is set off when screen turns on, RGB led is set off
or the alarm timed unblink happens. It sets on in each case vk
blink on is called.