fpf: squeeze control: squeeze to swipe functionality
Idea and first implementation by @tbalden

Addition of squeez to swipe: if set on, short squeeze
while screen on is taken over by an emulated swipe
scrolling on the touchscreen input device. Other
events while the emulated events are synced are filtered
out with a new input handler.
An after-squeeze swipe_longcounter work is added to
have a limit for a second short squeeze or middle long squeeze
to behave differently. If counter didnt finish,
a double swipe happens on a short squeeze to speed
up screen scrolling. The scrolling method takes into account
how long ago was the last scroll squeeze, and accelerates
the speed of swipe if little time passed.
If a middle long squeeze happens before counter finishing
there's a direction turning in scrolling.

In touchscreen input handler a new direction detection was
added, based on last vertical real touchscreen movement
the direction of the emulated scrolls is overwritten.

swipe_longcounter signals with a short vibration if enabled
when the time for a double speed short squeeze or direction
turn is over. After that middle long squeeze will mean
screen of if squeeze_sleep is enabled, or a direction turn.
Short squeeze means a normal single swipe again, and the
process starts from the beginning.

sysfs: fpf/squeeze_swipe and fpf/squeeze_swipe_vibration