vibrator: qpnp-haptic: notification booster feature
original idea and implementation by @tbalden
The haptic driver is extended to boost voltage when
timed device is being used for haptic feedback and
the time is over a certain limit. The limit should be
set to a value that generally is in line with notification
vibration lengths, and not keyboard/squeeze haptic feedbacks.

It's bound to screen off state also. It will reset voltage
to the originally set level so that all other haptic feedback
stays normal level.