fpf: KAD: fix KAD on charger only mode + BLN on, off charger no BLN
is_kernel_ambient_display() symbol will now return false if KAD
is on charger only mode and no charging is happening. This is used
in LED driver to trigger BLN only of KAD is not active, otherwise
screen waking of KAD will stop BLN anyway, so meaningless to start
it. But in this special case, KAD is not active off the charger.