uci: modification time check for sys cfg to avoid parse old data
check whether the sys file being parsed is over 3 seconds old
of inode mtime. That could mean the Companion app is not yet
running or terminated running abruptly, leaving important sys
state uci params in wrong state (after a reboot for example).
If so set outdated status, need of parse, and broadcast sys
params are to be reread with default values.

This avoids for example being stuck with blocked power
button because of proximity=1 status after a reboot.